requirements.txt
google-genai>=0.2.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
pathlib
json5
typing-extensions
